requirements.txt

fastapi==0.88.0
uvicorn
licensing
python-dotenv
timeloop